+++ browser.js
@@ -1,4 +1,4 @@
 /*
  * Wrap the exported function in a namespace for the browser.
  */
-module.exports = require('./src/index.js')
+export * from './src/index.js'